fix: 取消N+1次查询

+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/dto/stock/adjust/stock/StockAdjustProductDto.java
@@ -2,6 +2,7 @@ package com.lframework.xingyun.sc.dto.stock.adjust.stock;
 
 import com.lframework.starter.web.core.dto.BaseDto;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -77,5 +78,5 @@ public class StockAdjustProductDto implements BaseDto, Serializable {
   /**
    * 当前库存数量
    */
-  private Integer curStockNum;
+  private BigDecimal curStockNum;
 }
